# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartLibrary API",
        "description": "Dokumentasi API SmartLibrary",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/profiles": {
            "post": {
                "tags": [
                    "User & Profile"
                ],
                "summary": "Membuat atau memperbarui profil pengguna aktif",
                "operationId": "00abb5f24a20112b161d17e9a5c38a37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bio"
                                ],
                                "properties": {
                                    "bio": {
                                        "type": "string",
                                        "example": "Halo, saya Doni, seorang mahasiswa Teknologi Informasi."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profil berhasil disimpan"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profiles/{id}": {
            "get": {
                "tags": [
                    "User & Profile"
                ],
                "summary": "Melihat detail profil berdasarkan ID",
                "operationId": "bdbf886f973e16c2be2a83e76a804833",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari profil yang ingin dilihat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil ditemukan"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Profil tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reading-lists": {
            "get": {
                "tags": [
                    "Reading List"
                ],
                "summary": "Menampilkan semua daftar bacaan milik pengguna aktif",
                "operationId": "e2c932dd9a0916fd4e3a42a4e82a6f05",
                "responses": {
                    "200": {
                        "description": "Sukses mengambil data daftar bacaan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reading List"
                ],
                "summary": "Menambahkan buku ke dalam daftar bacaan",
                "operationId": "849149ea684c462a7a38a072d53a2daa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "want_to_read",
                                        "enum": [
                                            "want_to_read",
                                            "reading",
                                            "finished"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Buku berhasil dimasukkan ke daftar bacaan"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reading-lists/{id}": {
            "get": {
                "tags": [
                    "Reading List"
                ],
                "summary": "Melihat detail satu item daftar bacaan",
                "operationId": "b254fe5689c80eab16f18dc96814f9e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data item ditemukan"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reading List"
                ],
                "summary": "Mengubah status progres bacaan buku",
                "operationId": "88ca94ba734c3d28e81b4ef612435c73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "reading",
                                        "enum": [
                                            "want_to_read",
                                            "reading",
                                            "finished"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status bacaan berhasil diperbarui"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reading List"
                ],
                "summary": "Menghapus buku dari daftar bacaan",
                "operationId": "3787e2f4edfa32b735bd6a5608b00651",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item berhasil dihapus"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT kamu di sini (tanpa kata 'Bearer ')",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}